+++ NEWS.md
@@ -1,6 +1,7 @@
 # rsofun (development version)
 
-* Breaking change: biomee drivers' `init_cohort$init_n_cohorts` column is deprecated and should take the value 0.
+* Breaking change: biomee drivers' `init_cohort$init_n_cohorts` column has been phased out and must not be present in  
+drivers to protect against data corruption.
 
 # rsofun v5.0.0